Try platform specific code.
