+++ libgoal/lockedFile.go
@@ -18,20 +18,9 @@ package libgoal
 
 import (
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
-	"syscall"
 )
-
-// Platform-dependant locker implementation
-// How to extend
-// 1. Create two new files locker.go and locker_platform.go
-// 2. Put appropriate build tags
-// 3. Move unixLocker implementation and `newLockedFile` method to locker.go
-// 4. Implement platform-specific locker in locker_platform.go
-// 5. Ensure `newLockedFile` sets platform-specific locker
-//    so that lockedFile.read and lockedFile.write work correctly
 
 type locker interface {
 	tryRLock(fd *os.File) error
@@ -39,64 +28,11 @@ type locker interface {
 	unlock(fd *os.File) error
 }
 
-type unixLocker struct {
-	ofd         bool
-	setLock     int
-	setLockWait int
-	getLock     int
-}
-
-// makeUnixLocker create a unix file locker.
-// note that the desired way is to use the OFD locker, which locks on the file descriptor level.
-// falling back to the non-OFD lock would allow obtaining two locks by the same process. If this becomes
-// and issue, we might want to use flock, which wouldn't work across NFS.
-func makeUnixLocker() *unixLocker {
-	locker := &unixLocker{}
-	getlk := syscall.Flock_t{Type: syscall.F_RDLCK}
-	if err := syscall.FcntlFlock(0, 36 /*F_OFD_GETLK*/, &getlk); err == nil {
-		locker.ofd = true
-		// constants from /usr/include/bits/fcntl-linux.h
-		locker.setLock = 37     // F_OFD_SETLK
-		locker.setLockWait = 38 // F_OFD_SETLKW
-		locker.getLock = 36     // F_OFD_GETLK
-	} else {
-		locker.setLock = syscall.F_SETLK
-		locker.setLockWait = syscall.F_SETLKW
-		locker.getLock = syscall.F_GETLK
+func newLockedFile(path string) *lockedFile {
+	return &lockedFile{
+		path:   path,
+		locker: makeLocker(),
 	}
-	return locker
-}
-
-// the FcntlFlock has the most consistent behaviour across platforms,
-// and supports both local and network file systems.
-func (f *unixLocker) tryRLock(fd *os.File) error {
-	flock := &syscall.Flock_t{
-		Type:   syscall.F_RDLCK,
-		Whence: int16(io.SeekStart),
-		Start:  0,
-		Len:    0,
-	}
-	return syscall.FcntlFlock(fd.Fd(), f.setLockWait, flock)
-}
-
-func (f *unixLocker) tryLock(fd *os.File) error {
-	flock := &syscall.Flock_t{
-		Type:   syscall.F_WRLCK,
-		Whence: int16(io.SeekStart),
-		Start:  0,
-		Len:    0,
-	}
-	return syscall.FcntlFlock(fd.Fd(), f.setLockWait, flock)
-}
-
-func (f *unixLocker) unlock(fd *os.File) error {
-	flock := &syscall.Flock_t{
-		Type:   syscall.F_UNLCK,
-		Whence: int16(io.SeekStart),
-		Start:  0,
-		Len:    0,
-	}
-	return syscall.FcntlFlock(fd.Fd(), f.setLockWait, flock)
 }
 
 // lockedFile implementation
@@ -106,13 +42,6 @@ func (f *unixLocker) unlock(fd *os.File) error {
 type lockedFile struct {
 	path   string
 	locker locker
-}
-
-func newLockedFile(path string) *lockedFile {
-	return &lockedFile{
-		path:   path,
-		locker: makeUnixLocker(),
-	}
 }
 
 func (f *lockedFile) read() (bytes []byte, err error) {
